+++ server/tests-py/queries/graphql_query/permissions/inherited_roles/schema_teardown.sql
@@ -1,5 +1,0 @@
-type: run_sql
-args:
-  cascade: true
-  sql: |
-    DROP